# Optimize interop