--- a/Jint/Runtime/Interop/DelegateWrapper.cs
+++ b/Jint/Runtime/Interop/DelegateWrapper.cs
@@ -22,12 +22,12 @@ namespace Jint.Runtime.Interop
 
         public override JsValue Call(JsValue thisObject, JsValue[] jsArguments)
         {
-            var parameterInfos = _d.GetMethodInfo().GetParameters();
+            var parameterInfos = _d.Method.GetParameters();
 
             bool delegateContainsParamsArgument = false;
             foreach (var p in parameterInfos)
             {
-                if (p.HasAttribute<ParamArrayAttribute>())
+                if (Attribute.IsDefined(p, typeof(ParamArrayAttribute)))
                 {
                     delegateContainsParamsArgument = true;
                     break;
@@ -63,7 +63,7 @@ namespace Jint.Runtime.Interop
             // assign null to parameters not provided
             for (var i = jsArgumentsWithoutParamsCount; i < delegateNonParamsArgumentsCount; i++)
             {
-                if (parameterInfos[i].ParameterType.IsValueType())
+                if (parameterInfos[i].ParameterType.IsValueType)
                 {
                     parameters[i] = Activator.CreateInstance(parameterInfos[i].ParameterType);
                 }