… Jint/Runtime/Interop/NamespaceReference.cs
… Jint/Runtime/Interop/ObjectWrapper.cs
… Jint/Runtime/Interop/TypeReference.cs